pyproject.toml
[project]
name = "dl-cnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "opencv-contrib-python>=4.13.0.92",
    "optuna>=4.8.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "seaborn>=0.13.2",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
]
